feat: create MenuBuilderComponent
